save inner state
